Refactored ImageGalleryPanGestureDelegate
